Refactor App into modular components
