+++ src/App.jsx
@@ -1,80 +1,12 @@
-import React, { useMemo, useState, useEffect, useCallback, useRef, useLayoutEffect } from "react";
+import React, { useMemo, useState, useEffect, useCallback, useRef } from "react";
 import { ComposableMap, Geographies, Geography } from "react-simple-maps";
 import { geoEqualEarth, geoPath } from "d3-geo";
 import { ArrowLeftRight, Search, Info } from "lucide-react";
 
-// ---------------------------------------------------------------------------
-// UI primitives
-// ---------------------------------------------------------------------------
-// The design system in this project is intentionally tiny.  The components
-// below provide minimal styling wrappers so the rest of the file can focus on
-// application logic rather than className juggling.
-
-/** Generic card container with rounded corners + border. */
-const Card = ({ className = "", children }) => (
-  <div className={`bg-white border rounded-2xl ${className}`}>{children}</div>
-);
-
-/** Spacing wrapper used inside <Card /> elements. */
-const CardContent = ({ className = "", children }) => (
-  <div className={`p-4 sm:p-6 ${className}`}>{children}</div>
-);
-
-/**
- * Button primitive used everywhere in the UI.
- * The "variant" prop controls the color treatment.
- */
-const Button = ({
-  variant = "default",
-  className = "",
-  disabled,
-  onClick,
-  title,
-  children,
-}) => {
-  const baseClasses =
-    "inline-flex items-center justify-center gap-2 px-3 py-2 rounded-xl text-sm font-medium transition-all duration-200 border";
-
-  const variantClasses =
-    variant === "secondary"
-      ? "bg-slate-100 hover:bg-slate-200 border-slate-200"
-      : variant === "outline"
-      ? "bg-white hover:bg-slate-50 border-slate-300"
-      : "bg-slate-900 text-white hover:bg-black border-slate-900";
-
-  return (
-    <button
-      title={title}
-      disabled={disabled}
-      onClick={onClick}
-      className={`${baseClasses} ${variantClasses} disabled:opacity-60 disabled:cursor-not-allowed ${className}`}
-    >
-      {children}
-    </button>
-  );
-};
-
-/** Styled input used in search + filtering controls. */
-const Input = ({ className = "", ...props }) => (
-  <input
-    className={`w-full rounded-xl border border-slate-300 bg-white px-3 py-2 text-sm outline-none focus:ring-2 focus:ring-slate-300 ${className}`}
-    {...props}
-  />
-);
-
-/** Reusable select/dropdown control. */
-const Select = ({ value, onChange, children, className = "" }) => (
-  <select
-    value={value}
-    onChange={(event) => onChange(event.target.value)}
-    className={`rounded-xl border border-slate-300 bg-white px-3 py-2 text-sm outline-none focus:ring-2 focus:ring-slate-300 ${className}`}
-  >
-    {children}
-  </select>
-);
-
-/** Wrapper around <option> for readability. */
-const SelectItem = ({ value, children }) => <option value={value}>{children}</option>;
+import { Card, CardContent, Button, Input, Select, SelectItem } from "./components/ui";
+import { BoundedZoomableGroup } from "./hooks/useBoundedZoomPan";
+import { numberFmt, smallNumberFmt, legendFmt, pctFmt, percentFmt, fmtTime, relAB } from "./lib/formatters";
+import { StatRow } from "./components/StatRow";
 
 // ---------------------------------------------------------------------------
 // Map interaction helpers
@@ -96,359 +28,6 @@ const MAP_SELECTION_COLORS = {
 };
 
 const clampNumber = (value, min, max) => Math.min(Math.max(value, min), max);
-
-const identityConstrain = (position) => position;
-
-const useBoundedZoomPan = ({
-  center = [0, 0],
-  zoom = 1,
-  minZoom = 1,
-  maxZoom = 5,
-  zoomSensitivity = 0.025,
-  onZoomStart,
-  onZoomEnd,
-  onMoveStart,
-  onMove,
-  onMoveEnd,
-  disablePanning = false,
-  disableZooming = false,
-  width = 0,
-  height = 0,
-  projection,
-  constrain = identityConstrain,
-}) => {
-  const clampPosition = useCallback(
-    (pos) => {
-      const baseZoom = clampNumber(pos.zoom ?? zoom, minZoom, maxZoom);
-      const base = {
-        x: pos.x ?? 0,
-        y: pos.y ?? 0,
-        zoom: baseZoom,
-      };
-
-      try {
-        const constrained = constrain({ ...base });
-        if (!constrained) return base;
-        return {
-          x: Number.isFinite(constrained.x) ? constrained.x : base.x,
-          y: Number.isFinite(constrained.y) ? constrained.y : base.y,
-          zoom: clampNumber(constrained.zoom ?? base.zoom, minZoom, maxZoom),
-        };
-      } catch {
-        return base;
-      }
-    },
-    [constrain, maxZoom, minZoom, zoom]
-  );
-
-  const projectCenter = useCallback(
-    (coordinates, currentZoom) => {
-      if (!projection) {
-        return {
-          x: width / 2,
-          y: height / 2,
-        };
-      }
-
-      try {
-        const projected = projection(coordinates);
-        if (!Array.isArray(projected) || projected.length < 2) {
-          throw new Error("invalid projection");
-        }
-        const [px, py] = projected;
-        if (!Number.isFinite(px) || !Number.isFinite(py)) {
-          throw new Error("non-finite projection");
-        }
-        return {
-          x: width / 2 - px * currentZoom,
-          y: height / 2 - py * currentZoom,
-        };
-      } catch {
-        return {
-          x: width / 2,
-          y: height / 2,
-        };
-      }
-    },
-    [projection, width, height]
-  );
-
-  const [position, setPosition] = useState(() => {
-    const projected = projectCenter(center, zoom);
-    const clamped = clampPosition({ ...projected, zoom });
-    return {
-      x: clamped.x,
-      y: clamped.y,
-      last: [clamped.x, clamped.y],
-      zoom: clamped.zoom,
-      dragging: false,
-      zooming: false,
-    };
-  });
-
-  const elRef = useRef(null);
-  const point = useRef(null);
-  const wheelTimer = useRef(null);
-  const isPointerDown = useRef(false);
-  const pointerOrigin = useRef(null);
-
-  const getPointFromEvent = useCallback((event) => {
-    const svg = elRef.current?.closest("svg");
-    if (!svg) {
-      return { x: 0, y: 0 };
-    }
-
-    if (!point.current) {
-      point.current = svg.createSVGPoint();
-    }
-
-    if (event.targetTouches && event.targetTouches[0]) {
-      point.current.x = event.targetTouches[0].clientX;
-      point.current.y = event.targetTouches[0].clientY;
-    } else {
-      point.current.x = event.clientX;
-      point.current.y = event.clientY;
-    }
-
-    try {
-      const invertedMatrix = svg.getScreenCTM().inverse();
-      return point.current.matrixTransform(invertedMatrix);
-    } catch {
-      return { x: 0, y: 0 };
-    }
-  }, []);
-
-  const handlePointerDown = useCallback(
-    (event) => {
-      if (disablePanning) return;
-      const svg = elRef.current?.closest("svg");
-      if (!svg) return;
-
-      isPointerDown.current = true;
-      pointerOrigin.current = getPointFromEvent(event);
-
-      setPosition((current) => {
-        const next = { ...current, dragging: true };
-        if (onMoveStart) onMoveStart(event, next);
-        return next;
-      });
-    },
-    [disablePanning, getPointFromEvent, onMoveStart]
-  );
-
-  const handlePointerMove = useCallback(
-    (event) => {
-      if (!isPointerDown.current) return;
-      event.preventDefault();
-      const pointerPosition = getPointFromEvent(event);
-
-      setPosition((current) => {
-        if (!pointerOrigin.current) return current;
-        const raw = {
-          ...current,
-          x: current.last[0] + (pointerPosition.x - pointerOrigin.current.x),
-          y: current.last[1] + (pointerPosition.y - pointerOrigin.current.y),
-          dragging: true,
-        };
-        const clamped = clampPosition(raw);
-        const next = {
-          ...current,
-          x: clamped.x,
-          y: clamped.y,
-          zoom: clamped.zoom,
-          dragging: true,
-        };
-        if (onMove) onMove(event, next);
-        return next;
-      });
-    },
-    [clampPosition, getPointFromEvent, onMove]
-  );
-
-  const handlePointerUp = useCallback(
-    (event) => {
-      if (!isPointerDown.current) return;
-      isPointerDown.current = false;
-
-      setPosition((current) => {
-        const clamped = clampPosition(current);
-        const next = {
-          ...current,
-          x: clamped.x,
-          y: clamped.y,
-          zoom: clamped.zoom,
-          last: [clamped.x, clamped.y],
-          dragging: false,
-        };
-        if (onMoveEnd) onMoveEnd(event, next);
-        return next;
-      });
-    },
-    [clampPosition, onMoveEnd]
-  );
-
-  const handleWheel = useCallback(
-    (event) => {
-      if (!event.ctrlKey || disableZooming) return;
-      event.preventDefault();
-
-      const speed = event.deltaY * zoomSensitivity;
-
-      setPosition((current) => {
-        const newZoom = clampNumber(current.zoom - speed, minZoom, maxZoom);
-        const pointerPosition = getPointFromEvent(event);
-
-        const rawX = (current.x - pointerPosition.x) * (newZoom / current.zoom) + pointerPosition.x;
-        const rawY = (current.y - pointerPosition.y) * (newZoom / current.zoom) + pointerPosition.y;
-
-        let next = {
-          ...current,
-          x: rawX,
-          y: rawY,
-          zoom: newZoom,
-          zooming: true,
-          last: [rawX, rawY],
-        };
-
-        next = { ...next, ...clampPosition(next) };
-        next.last = [next.x, next.y];
-
-        window.clearTimeout(wheelTimer.current);
-        wheelTimer.current = window.setTimeout(() => {
-          setPosition((finalState) => ({ ...finalState, zooming: false }));
-          if (onZoomEnd) onZoomEnd(event, next);
-        }, 66);
-
-        if (onZoomStart) onZoomStart(event, next);
-
-        return next;
-      });
-    },
-    [clampPosition, disableZooming, getPointFromEvent, maxZoom, minZoom, onZoomEnd, onZoomStart, zoomSensitivity]
-  );
-
-  useLayoutEffect(() => {
-    const svg = elRef.current?.closest("svg");
-    if (!svg) return undefined;
-
-    const down = (event) => handlePointerDown(event);
-    const move = (event) => handlePointerMove(event);
-    const up = (event) => handlePointerUp(event);
-
-    svg.addEventListener("wheel", handleWheel, { passive: false });
-
-    if (window.PointerEvent) {
-      svg.addEventListener("pointerdown", down);
-      svg.addEventListener("pointermove", move, { passive: false });
-      svg.addEventListener("pointerup", up);
-      svg.addEventListener("pointerleave", up);
-    } else {
-      svg.addEventListener("mousedown", down);
-      svg.addEventListener("mousemove", move);
-      svg.addEventListener("mouseup", up);
-      svg.addEventListener("mouseleave", up);
-      svg.addEventListener("touchstart", down);
-      svg.addEventListener("touchmove", move, { passive: false });
-      svg.addEventListener("touchend", up);
-    }
-
-    return () => {
-      svg.removeEventListener("wheel", handleWheel);
-
-      if (window.PointerEvent) {
-        svg.removeEventListener("pointerdown", down);
-        svg.removeEventListener("pointermove", move);
-        svg.removeEventListener("pointerup", up);
-        svg.removeEventListener("pointerleave", up);
-      } else {
-        svg.removeEventListener("mousedown", down);
-        svg.removeEventListener("mousemove", move);
-        svg.removeEventListener("mouseup", up);
-        svg.removeEventListener("mouseleave", up);
-        svg.removeEventListener("touchstart", down);
-        svg.removeEventListener("touchmove", move);
-        svg.removeEventListener("touchend", up);
-      }
-    };
-  }, [handlePointerDown, handlePointerMove, handlePointerUp, handleWheel]);
-
-  useEffect(() => {
-    setPosition((current) => {
-      const updated = clampPosition({ ...current, zoom });
-      return {
-        ...current,
-        x: updated.x,
-        y: updated.y,
-        zoom: updated.zoom,
-        last: [updated.x, updated.y],
-      };
-    });
-  }, [clampPosition, zoom]);
-
-  useEffect(() => {
-    setPosition((current) => {
-      const projected = projectCenter(center, current.zoom);
-      const updated = clampPosition({ ...current, ...projected });
-      return {
-        ...current,
-        x: updated.x,
-        y: updated.y,
-        zoom: updated.zoom,
-        last: [updated.x, updated.y],
-      };
-    });
-  }, [center, clampPosition, projectCenter]);
-
-  useEffect(() => {
-    setPosition((current) => {
-      const updated = clampPosition(current);
-      return {
-        ...current,
-        x: updated.x,
-        y: updated.y,
-        zoom: updated.zoom,
-        last: [updated.x, updated.y],
-      };
-    });
-  }, [clampPosition]);
-
-  useEffect(() => () => {
-    if (wheelTimer.current) {
-      window.clearTimeout(wheelTimer.current);
-    }
-  }, []);
-
-  return {
-    elRef,
-    position,
-    transformString: `translate(${position.x} ${position.y}) scale(${position.zoom})`,
-  };
-};
-
-const BoundedZoomableGroup = ({
-  children,
-  render,
-  className = "",
-  width,
-  height,
-  projection,
-  constrain,
-  ...rest
-}) => {
-  const { elRef, position, transformString } = useBoundedZoomPan({
-    width,
-    height,
-    projection,
-    constrain,
-    ...rest,
-  });
-
-  return (
-    <g ref={elRef} className={`rsm-zoomable-group ${className}`}>
-      {render ? render(position) : <g transform={transformString}>{children}</g>}
-    </g>
-  );
-};
 
 // ---------------------------------------------------------------------------
 // Constants & helper utilities
@@ -535,221 +114,6 @@ const whiteBlue = (t) => {
 /** Generate `n` color stops using the gradient above. */
 const whiteBluePalette = (n) =>
   Array.from({ length: n }, (_, index) => whiteBlue(n === 1 ? 1 : index / (n - 1)));
-
-/**
- * Human friendly number formatter used for big magnitudes (thousands, millions…).
- * Returns an em dash for nullish / empty values so the table stays legible.
- */
-const numberFmt = (value) => {
-  if (value === null || value === undefined || value === "") return "—";
-
-  const numeric = Number(value);
-  if (!Number.isFinite(numeric)) return String(value);
-
-  const absValue = Math.abs(numeric);
-  if (absValue >= 1e12) return (numeric / 1e12).toFixed(1) + "T";
-  if (absValue >= 1e9) return (numeric / 1e9).toFixed(1) + "B";
-  if (absValue >= 1e6) return (numeric / 1e6).toFixed(1) + "M";
-  if (absValue >= 1e3) return (numeric / 1e3).toFixed(1) + "K";
-
-  return String(numeric);
-};
-
-/** Helper used by smallNumberFmt + pctFmt to avoid floating point noise. */
-const truncateTo = (value, digits = 1) => Math.trunc(Number(value) * 10 ** digits) / 10 ** digits;
-
-/**
- * Produce compact decimal numbers while preserving trailing zeros only when
- * necessary.  Useful for metrics like life expectancy.
- */
-const smallNumberFmt = (value, digits = 1) => {
-  if (value === null || value === undefined || value === "" || Number.isNaN(Number(value))) {
-    return "—";
-  }
-
-  return truncateTo(Number(value), digits)
-    .toFixed(digits)
-    .replace(/\.0+$/, "")
-    .replace(/\.$/, "");
-};
-
-/** Format numbers specifically for use in the legend labels. */
-const legendFmt = (value) => {
-  const numeric = Number(value);
-  if (!Number.isFinite(numeric)) return "—";
-
-  return Math.abs(numeric) >= 1e3 ? numberFmt(numeric) : smallNumberFmt(numeric, 1);
-};
-
-/** Format fractional values (0.2) as signed percentages (+20.0%). */
-const pctFmt = (value) => {
-  if (value === null || value === undefined || Number.isNaN(value)) return "—";
-
-  const sign = value >= 0 ? "+" : "";
-  return sign + truncateTo(value * 100, 1).toFixed(1) + "%";
-};
-
-/** Format absolute percentage values (20) with consistent precision. */
-const percentFmt = (value, digits = 1) => {
-  if (value === null || value === undefined || value === "") return "—";
-
-  const numeric = Number(value);
-  if (!Number.isFinite(numeric)) return "—";
-
-  return truncateTo(numeric, digits)
-    .toFixed(digits)
-    .replace(/\.0+$/, "")
-    .replace(/\.$/, "")
-    .concat("%");
-};
-
-/** Present timestamps consistently with local time + 24h clock. */
-const fmtTime = (timestamp) => {
-  try {
-    return new Date(timestamp).toLocaleString(undefined, {
-      year: "numeric",
-      month: "short",
-      day: "2-digit",
-      hour: "2-digit",
-      minute: "2-digit",
-      hour12: false,
-    });
-  } catch {
-    return String(timestamp);
-  }
-};
-
-/**
- * Small component that renders the absolute + percentage difference between
- * two numeric values (used in the comparison table).
- */
-const DiffCell = ({ a, b, suffix = "" }) => {
-  if (a == null || b == null || a === "" || b === "") return <span>—</span>;
-
-  const diff = b - a;
-  const pct = a !== 0 ? diff / a : null;
-  const sign = diff > 0 ? "+" : diff < 0 ? "" : "";
-  const abs = Math.abs(diff);
-  const display = abs >= 1000 ? numberFmt(abs) : smallNumberFmt(abs, 1);
-  const trendClass =
-    diff > 0 ? "text-emerald-600" : diff < 0 ? "text-rose-600" : "text-slate-500";
-
-  return (
-    <div className="text-sm text-right">
-      <div className={`font-medium ${trendClass}`}>
-        {sign}
-        {display}
-        {suffix}
-      </div>
-      <div className="text-xs text-slate-500 text-right">{pctFmt(pct)}</div>
-    </div>
-  );
-};
-
-/** Determine whether country A or B has the larger numeric value. */
-const relAB = (a, b) => {
-  if (a == null || a === "" || b == null || b === "") return "na";
-
-  const an = Number(a);
-  const bn = Number(b);
-  if (!Number.isFinite(an) || !Number.isFinite(bn)) return "na";
-
-  if (an === bn) return "tie";
-  return an > bn ? "A" : "B";
-};
-
-const StatRow = ({
-  label,
-  field,
-  dataA,
-  dataB,
-  fmt = (value) => value,
-  diffSuffix = "",
-  defaultYear,
-}) => {
-  const seriesA = dataA?.__series?.[field] || {};
-  const seriesB = dataB?.__series?.[field] || {};
-
-  const yearsA = Object.keys(seriesA).map(Number).filter(Number.isFinite);
-  const yearsB = Object.keys(seriesB).map(Number).filter(Number.isFinite);
-
-  // Prefer the latest year available in both datasets. If there is no common
-  // year, fall back to the most recent year seen in either country.
-  const commonYears = yearsA.filter((year) => yearsB.includes(year));
-  let rowYear = null;
-  if (commonYears.length) {
-    rowYear = Math.max(...commonYears);
-  } else if (yearsA.length || yearsB.length) {
-    rowYear = Math.max(...(yearsA.length ? yearsA : [-Infinity]), ...(yearsB.length ? yearsB : [-Infinity]));
-  }
-
-  const valueA = rowYear != null ? seriesA[rowYear] : null;
-  const valueB = rowYear != null ? seriesB[rowYear] : null;
-  const relationship = relAB(valueA, valueB);
-
-  const icon = (kind) => (
-    <span
-      className="ml-2 inline-flex items-center rounded px-1.5 py-0.5 ring-1 shrink-0"
-      style={{
-        backgroundColor: kind === "higher" ? "rgba(16,185,129,0.12)" : "rgba(244,63,94,0.10)",
-        color: kind === "higher" ? "rgb(5,122,85)" : "rgb(190,18,60)",
-        borderColor: kind === "higher" ? "rgba(16,185,129,0.35)" : "rgba(244,63,94,0.35)",
-      }}
-      title={kind === "higher" ? "Higher" : "Lower"}
-    >
-      <span aria-hidden="true" className="text-[10px] leading-none">
-        {kind === "higher" ? "▲" : "▼"}
-      </span>
-      <span className="sr-only">{kind === "higher" ? "Higher" : "Lower"}</span>
-    </span>
-  );
-
-  const metricLabel = (
-    <div className="flex items-start justify-between gap-2 min-w-0">
-      <span className="flex-1 min-w-0 whitespace-normal break-words">{label}</span>
-      {rowYear && defaultYear && rowYear < defaultYear ? (
-        <span
-          className="text-[10px] text-amber-600 shrink-0 whitespace-nowrap tabular-nums text-right"
-          title={`Older data: ${rowYear} (default ${defaultYear})`}
-        >
-          ({rowYear})
-        </span>
-      ) : null}
-    </div>
-  );
-
-  const highlightA = relationship === "A" ? "bg-emerald-50/30" : relationship === "B" ? "bg-rose-50/30" : "";
-  const highlightB = relationship === "B" ? "bg-emerald-50/30" : relationship === "A" ? "bg-rose-50/30" : "";
-  const displayA = fmt(valueA);
-  const displayB = fmt(valueB);
-
-  return (
-    <tr className="border-b last:border-0">
-      <td className="py-2 pr-3 text-xs sm:text-sm text-slate-500 align-top">{metricLabel}</td>
-      <td className={`py-2 pr-2 font-medium ${highlightA}`}>
-        <div className="flex items-center gap-1 min-w-0">
-          <span className="truncate" title={displayA}>
-            {displayA}
-          </span>
-          {relationship === "A" && icon("higher")}
-          {relationship === "B" && icon("lower")}
-        </div>
-      </td>
-      <td className={`py-2 pr-2 font-medium ${highlightB}`}>
-        <div className="flex items-center gap-1 min-w-0">
-          <span className="truncate" title={displayB}>
-            {displayB}
-          </span>
-          {relationship === "B" && icon("higher")}
-          {relationship === "A" && icon("lower")}
-        </div>
-      </td>
-      <td className="py-2 pl-2 align-top">
-        <DiffCell a={valueA} b={valueB} suffix={diffSuffix} />
-      </td>
-    </tr>
-  );
-};
 
 const SearchBox = ({ placeholder, value, onChange }) => (
   <div className="relative w-full">
@@ -1253,16 +617,6 @@ export default function App() {
     return findMetricConfig(field)?.label || field;
   };
 
-  const fmtFor = (field) => {
-    if (!field) return numberFmt;
-    return findMetricConfig(field)?.fmt || numberFmt;
-  };
-
-  const diffSuffixFor = (field) => {
-    if (!field) return "";
-    return findMetricConfig(field)?.diffSuffix || "";
-  };
-
   const metricGroups = useMemo(() => {
     if (!metrics.length) return [];
 
@@ -1418,7 +772,7 @@ export default function App() {
       const t = (value - linearStats.min) / (linearStats.max - linearStats.min);
       return whiteBlue(t);
     },
-    [colorScaleMode, linearStats, valueStats]
+    [colorMetric, colorScaleMode, linearStats, valueStats]
   );
 
   const filtered = useMemo(() => {
